Use Ampere instead of IU for torque, and change input/return type of some methods

+++ march_hardware/include/march_hardware/ethercat/slave.h
@@ -31,7 +31,7 @@ public:
 
   ~Slave() noexcept override = default;
 
-  uint16_t getSlaveIndex() const
+  int getSlaveIndex() const
   {
     return this->slave_index_;
   }
