+++ march_hardware/src/joint.cpp
@@ -141,7 +141,7 @@ double Joint::getAbsolutePosition() const
   return this->absolute_position_;
 }
 
-void Joint::actuateTorque(int16_t target_torque)
+void Joint::actuateTorque(double target_torque)
 {
   if (!this->canActuate())
   {
@@ -151,7 +151,7 @@ void Joint::actuateTorque(int16_t target_torque)
   this->controller_->actuateTorque(target_torque);
 }
 
-int16_t Joint::getTorque()
+double Joint::getTorque()
 {
   return this->controller_->getTorque();
 }
