+++ march_hardware/test/mocks/mock_motor_controller.h
@@ -18,7 +18,7 @@ public:
   MOCK_METHOD0(prepareActuation, void());
   MOCK_CONST_METHOD0(getActuationMode, march::ActuationMode());
 
-  MOCK_CONST_METHOD0(getSlaveIndex, uint16_t());
+  MOCK_CONST_METHOD0(getSlaveIndex, int());
   MOCK_CONST_METHOD0(getIncrementalMorePrecise, bool());
   MOCK_METHOD0(getStates, march::MotorControllerStates&());
 
@@ -30,10 +30,10 @@ public:
   MOCK_METHOD0(getMotorControllerVoltage, float());
   MOCK_METHOD0(getMotorVoltage, float());
   MOCK_METHOD0(getMotorCurrent, float());
-  MOCK_METHOD0(getTorque, int16_t());
+  MOCK_METHOD0(getTorque, double());
 
   MOCK_METHOD1(actuateRad, void(double));
-  MOCK_METHOD1(actuateTorque, void(int16_t));
+  MOCK_METHOD1(actuateTorque, void(double));
 
   MOCK_METHOD1(initialize, bool(int cycle_time));
   MOCK_METHOD0(reset, void());
